+++ specs/reporter-factory.spec.php
@@ -3,6 +3,7 @@ use Evenement\EventEmitter;
 use Peridot\Configuration;
 use Peridot\Core\Suite;
 use Peridot\Reporter\AnonymousReporter;
+use Peridot\Reporter\CompositeReporter;
 use Peridot\Reporter\ReporterFactory;
 use Peridot\Reporter\SpecReporter;
 use Peridot\Runner\Runner;
@@ -48,6 +49,55 @@ describe('ReporterFactory', function() {
                 try {
                     $this->factory->create('nope');
                 } catch (RuntimeException $e) {
+                    $exception = $e;
+                }
+                assert(!is_null($exception), 'exception should have been thrown');
+            });
+        });
+    });
+
+    describe('->createComposite()', function() {
+        context('using valid reporter names', function() {
+            it('should return a composite of the named reporters', function() {
+                $this->factory->register('spec2', 'desc', function($reporter) {});
+                $reporter = $this->factory->createComposite(['spec', 'spec2']);
+                $reporters = $reporter->getReporters();
+                assert($reporter instanceof CompositeReporter, 'should create CompositeReporter');
+                assert($reporters[0] instanceof SpecReporter, 'first reporter should be a SpecReporter');
+            });
+        });
+
+        context('using valid names with invalid factories', function() {
+            it('should throw an exception', function() {
+                $this->factory->register('nope', 'doesnt work', 'Not\A\Class');
+                $exception = null;
+                try {
+                    $this->factory->createComposite(['spec', 'nope']);
+                } catch (RuntimeException $e) {
+                    $exception = $e;
+                }
+                assert(!is_null($exception), 'exception should have been thrown');
+            });
+        });
+
+        context('using invalid names', function() {
+            it('should throw an exception', function() {
+                $exception = null;
+                try {
+                    $this->factory->createComposite(['spec', 'nope']);
+                } catch (RuntimeException $e) {
+                    $exception = $e;
+                }
+                assert(!is_null($exception), 'exception should have been thrown');
+            });
+        });
+
+        context('using an empty name list', function() {
+            it('should throw an exception', function() {
+                $exception = null;
+                try {
+                    $this->factory->createComposite([]);
+                } catch (InvalidArgumentException $e) {
                     $exception = $e;
                 }
                 assert(!is_null($exception), 'exception should have been thrown');