+++ src/Reporter/ReporterFactory.php
@@ -4,6 +4,7 @@ namespace Peridot\Reporter;
 use Evenement\EventEmitterInterface;
 use Peridot\Configuration;
 use Peridot\Core\HasEventEmitterTrait;
+use Symfony\Component\Console\Output\BufferedOutput;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
@@ -62,19 +63,32 @@ class ReporterFactory
      */
     public function create($name)
     {
-        $factory = $this->getReporterFactory($name);
+        return $this->createWithOutput($this->output, $name);
+    }
 
-        $isClass = is_string($factory) && class_exists($factory);
-        
-        if ($isClass) {
-            return new $factory($this->configuration, $this->output, $this->eventEmitter);
+    /**
+     * Return an instance of the named reporter
+     *
+     * @param $name
+     * @return \Peridot\Reporter\AbstractBaseReporter
+     */
+    public function createComposite(array $names)
+    {
+        if (empty($names)) {
+            throw new \InvalidArgumentException('Reporter names cannot be empty.');
         }
 
-        if (is_callable($factory)) {
-            return new AnonymousReporter($factory, $this->configuration, $this->output, $this->eventEmitter);
-        }
-
-        throw new \RuntimeException("Reporter class could not be created");
+        return new CompositeReporter(
+            array_merge(
+                [$this->createWithOutput($this->output, array_shift($names))],
+                array_map(function ($name) {
+                    return $this->createWithOutput(new BufferedOutput(), $name);
+                }, $names)
+            ),
+            $this->configuration,
+            $this->output,
+            $this->eventEmitter
+        );
     }
 
     /**
@@ -125,5 +139,21 @@ class ReporterFactory
     public function getReporters()
     {
         return $this->reporters;
+    }
+
+    private function createWithOutput(OutputInterface $output, $name)
+    {
+        $factory = $this->getReporterFactory($name);
+        $isClass = is_string($factory) && class_exists($factory);
+
+        if ($isClass) {
+            return new $factory($this->configuration, $output, $this->eventEmitter);
+        }
+
+        if (is_callable($factory)) {
+            return new AnonymousReporter($factory, $this->configuration, $output, $this->eventEmitter);
+        }
+
+        throw new \RuntimeException("Reporter class could not be created");
     }
 }